📦 NEW: #23 Update page navigation concept to make /search the main page
* Replace index.svelte with search.svelte
* Remove old index.svelte
* Fix tests
* Fix routing
* Add redirect for "/search" to "/"
