+++ src/components/FlagDetails/Header.svelte
@@ -160,7 +160,7 @@
 
 <header data-cy-flag-header data-state={state}>
   <div class="header_inner">
-    <a href="./search" class="goBack">
+    <a href="/" class="goBack">
       <Icon
         icon="arrow-back"
         scale="var(--spacing-large)"
