+++ src/routes/flag/index.svelte
@@ -1,6 +1,6 @@
 <script context="module">
   export async function preload(page, session) {
-    return this.redirect(308, "/search");
+    return this.redirect(308, "/");
   }
 </script>
